Next: Move to `app/` directory from `pages/` directory - Need to movce the APIs to app dir
