+++ app/components/AnimatedBars.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { animate } from "motion";
 import { useEffect } from "react";
