+++ app/components/Button.tsx
@@ -1,3 +1,4 @@
+"use client";
 import { twMerge } from "tailwind-merge";
 
 export type ButtonPropsType = {
