+++ app/components/CardLinkButton.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { ArrowUpRightIcon } from "@heroicons/react/24/outline";
 import { twMerge } from "tailwind-merge";
