+++ app/components/Footer.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { Player } from "@lottiefiles/react-lottie-player";
 import Heart from "@public/emoji/heart.json";
