+++ app/components/Navbar.tsx
@@ -1,9 +1,9 @@
-import Button from "@components/Button";
+"use client";
 import fetcher from "@lib/fetcher";
-import Head from "next/head";
 import Link from "next/link";
-import { useRouter } from "next/router";
+import { usePathname } from "next/navigation";
 import useSWRImmutable from "swr/immutable";
+import Button from "./Button";
 
 type CommitSha = {
   sha: string;
@@ -17,14 +17,13 @@ type LinkWrapperProps = {
 };
 
 export default function Navbar() {
-  const { pathname } = useRouter();
+  const pathname = usePathname();
 
   const { data } = useSWRImmutable<CommitSha>(
     "https://api.github.com/repos/chinmaykunkikar/portfolio/commits/next",
     fetcher
   );
   const shortSha = data?.sha.slice(0, 7);
-  // const shortSha = "f77705a";
 
   function LinkWrapper({
     children,
@@ -44,11 +43,6 @@ export default function Navbar() {
 
   return (
     <>
-      <Head>
-        <title>Chinmay Kunkikar</title>
-        <meta name="description" content="Portfolio of Chinmay Kunkikar" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
       <div className="flex w-screen flex-col items-center justify-between gap-4 px-12 py-8 text-neutral-700 md:px-20 lg:flex-row lg:gap-0 lg:px-32">
         <span className="text-xl">
           <h3 className="cursor-default	select-none	font-mono">
