+++ app/components/drawables/node.tsx
