+++ app/components/drawables/terminal.tsx
