+++ app/components/widgets/PhotoWidget.tsx
@@ -1,3 +1,4 @@
+"use client";
 import Card from "@components/Card";
 import Me from "@public/me/me.jpeg";
 import Image from "next/image";
