+++ app/components/widgets/ResumeWidget.tsx
@@ -1,3 +1,4 @@
+"use client";
 import Card from "@components/Card";
 import CardLinkButton from "@components/CardLinkButton";
 import ResumeIcon from "@components/drawables/resume";
