+++ app/components/widgets/SpotifyWidget.tsx
@@ -1,8 +1,9 @@
+"use client";
 import useSWR from "swr";
 
 import AnimatedBars from "@components/AnimatedBars";
 import Card from "@components/Card";
-import SpotifyLogo from "@drawables/spotify";
+import SpotifyLogo from "@components/drawables/spotify";
 import fetcher from "@lib/fetcher";
 import { NowPlayingSong } from "@pages/api/now-playing";
 import { RecentlyPlayedSong } from "@pages/api/recently-played";
