+++ app/components/widgets/TimeWidget.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { ClockIcon } from "@heroicons/react/24/outline";
 import { useEffect, useState } from "react";
