+++ app/components/widgets/VideWidget.tsx
@@ -1,3 +1,4 @@
+"use client";
 import Card from "@components/Card";
 import CardLinkButton from "@components/CardLinkButton";
 import { ArrowRightIcon, ArrowUpRightIcon } from "@heroicons/react/24/outline";
